chore:move Service to package
